Splitting NonBlockingStatsDClient into its blocking and non-blocking parts w/ tests.
