+++ src/main/java/com/timgroup/statsd/BlockingStatsDClient.java
@@ -1,0 +1,654 @@
+package com.timgroup.statsd;
+
+import java.io.IOException;
+import java.net.InetAddress;
+import java.net.InetSocketAddress;
+import java.net.UnknownHostException;
+import java.nio.ByteBuffer;
+import java.nio.channels.DatagramChannel;
+import java.nio.charset.Charset;
+import java.text.DecimalFormat;
+import java.text.DecimalFormatSymbols;
+import java.text.NumberFormat;
+import java.util.Locale;
+import java.util.concurrent.Callable;
+
+/**
+ * A simple StatsD client implementation facilitating metrics recording.
+ *
+ * <p>Upon instantiation, this client will establish a socket connection to a StatsD instance
+ * running on the specified host and port. Metrics are then sent over this connection as they are
+ * received by the client.
+ * </p>
+ *
+ * <p>Three key methods are provided for the submission of data-points for the application under
+ * scrutiny:
+ * <ul>
+ *   <li>{@link #incrementCounter} - adds one to the value of the specified named counter</li>
+ *   <li>{@link #recordGaugeValue} - records the latest fixed value for the specified named gauge</li>
+ *   <li>{@link #recordExecutionTime} - records an execution time in milliseconds for the specified named operation</li>
+ *   <li>{@link #recordHistogramValue} - records a value, to be tracked with average, maximum, and percentiles</li>
+ *   <li>{@link #recordEvent} - records an event</li>
+ *   <li>{@link #recordSetValue} - records a value in a set</li>
+ * </ul>
+ * From the perspective of the application, these methods are non-blocking, with the resulting
+ * IO operations being carried out in a separate thread. Furthermore, these methods are guaranteed
+ * not to throw an exception which may disrupt application execution.
+ * </p>
+ *
+ * <p>As part of a clean system shutdown, the {@link #stop()} method should be invoked
+ * on any StatsD clients.</p>
+ *
+ * @author Tom Denley
+ *
+ */
+public class BlockingStatsDClient implements StatsDClient {
+
+    private static final int PACKET_SIZE_BYTES = 1500;
+
+    /**
+     * Because NumberFormat is not thread-safe we cannot share instances across threads. Use a ThreadLocal to
+     * create one pre thread as this seems to offer a significant performance improvement over creating one per-thread:
+     * http://stackoverflow.com/a/1285297/2648
+     * https://github.com/indeedeng/java-dogstatsd-client/issues/4
+     */
+    private static final ThreadLocal<NumberFormat> NUMBER_FORMATTERS = new ThreadLocal<NumberFormat>() {
+        @Override
+        protected NumberFormat initialValue() {
+
+            // Always create the formatter for the US locale in order to avoid this bug:
+            // https://github.com/indeedeng/java-dogstatsd-client/issues/3
+            final NumberFormat numberFormatter = NumberFormat.getInstance(Locale.US);
+            numberFormatter.setGroupingUsed(false);
+            numberFormatter.setMaximumFractionDigits(6);
+
+            // we need to specify a value for Double.NaN that is recognized by dogStatsD
+            if (numberFormatter instanceof DecimalFormat) { // better safe than a runtime error
+                final DecimalFormat decimalFormat = (DecimalFormat) numberFormatter;
+                final DecimalFormatSymbols symbols = decimalFormat.getDecimalFormatSymbols();
+                symbols.setNaN("NaN");
+                decimalFormat.setDecimalFormatSymbols(symbols);
+            }
+
+            return numberFormatter;
+        }
+    };
+
+    private final String prefix;
+    private final DatagramChannel clientChannel;
+    protected final StatsDClientErrorHandler handler;
+    private final String constantTagsRendered;
+    private final Callable<InetSocketAddress> addressLookup;
+    
+    /**
+     * Create a new StatsD client communicating with a StatsD instance on the
+     * specified host and port. All messages send via this client will have
+     * their keys prefixed with the specified string. The new client will
+     * attempt to open a connection to the StatsD server immediately upon
+     * instantiation, and may throw an exception if that a connection cannot
+     * be established. Once a client has been instantiated in this way, all
+     * exceptions thrown during subsequent usage are consumed, guaranteeing
+     * that failures in metrics will not affect normal code execution.
+     *
+     * @param prefix
+     *     the prefix to apply to keys sent via this client
+     * @param hostname
+     *     the host name of the targeted StatsD server
+     * @param port
+     *     the port of the targeted StatsD server
+     * @throws StatsDClientException
+     *     if the client could not be started
+     */
+    public BlockingStatsDClient(final String prefix, final String hostname, final int port) throws StatsDClientException {
+        this(prefix, hostname, port, null, null);
+    }
+
+    /**
+     * Create a new StatsD client communicating with a StatsD instance on the
+     * specified host and port. All messages send via this client will have
+     * their keys prefixed with the specified string. The new client will
+     * attempt to open a connection to the StatsD server immediately upon
+     * instantiation, and may throw an exception if that a connection cannot
+     * be established. Once a client has been instantiated in this way, all
+     * exceptions thrown during subsequent usage are consumed, guaranteeing
+     * that failures in metrics will not affect normal code execution.
+     *
+     * @param prefix
+     *     the prefix to apply to keys sent via this client
+     * @param hostname
+     *     the host name of the targeted StatsD server
+     * @param port
+     *     the port of the targeted StatsD server
+     * @param constantTags
+     *     tags to be added to all content sent
+     * @throws StatsDClientException
+     *     if the client could not be started
+     */
+    public BlockingStatsDClient(final String prefix, final String hostname, final int port, final String... constantTags) throws StatsDClientException {
+        this(prefix, hostname, port, constantTags, null);
+    }
+
+    /**
+     * Create a new StatsD client communicating with a StatsD instance on the
+     * specified host and port. All messages send via this client will have
+     * their keys prefixed with the specified string. The new client will
+     * attempt to open a connection to the StatsD server immediately upon
+     * instantiation, and may throw an exception if that a connection cannot
+     * be established. Once a client has been instantiated in this way, all
+     * exceptions thrown during subsequent usage are passed to the specified
+     * handler and then consumed, guaranteeing that failures in metrics will
+     * not affect normal code execution.
+     *
+     * @param prefix
+     *     the prefix to apply to keys sent via this client
+     * @param hostname
+     *     the host name of the targeted StatsD server
+     * @param port
+     *     the port of the targeted StatsD server
+     * @param constantTags
+     *     tags to be added to all content sent
+     * @param errorHandler
+     *     handler to use when an exception occurs during usage, may be null to indicate noop
+     * @throws StatsDClientException
+     *     if the client could not be started
+     */
+    public BlockingStatsDClient(final String prefix, final String hostname, final int port, final String[] constantTags,
+                                   final StatsDClientErrorHandler errorHandler) throws StatsDClientException {
+        this(prefix, constantTags, errorHandler, staticStatsDAddressResolution(hostname, port));
+    }
+
+    /**
+     * Create a new StatsD client communicating with a StatsD instance on the
+     * specified host and port. All messages send via this client will have
+     * their keys prefixed with the specified string. The new client will
+     * attempt to open a connection to the StatsD server immediately upon
+     * instantiation, and may throw an exception if that a connection cannot
+     * be established. Once a client has been instantiated in this way, all
+     * exceptions thrown during subsequent usage are passed to the specified
+     * handler and then consumed, guaranteeing that failures in metrics will
+     * not affect normal code execution.
+     *
+     * @param prefix
+     *     the prefix to apply to keys sent via this client
+     * @param constantTags
+     *     tags to be added to all content sent
+     * @param errorHandler
+     *     handler to use when an exception occurs during usage, may be null to indicate noop
+     * @param addressLookup
+     *     yields the IP address and socket of the StatsD server
+     * @throws StatsDClientException
+     *     if the client could not be started
+     */
+    public BlockingStatsDClient(final String prefix, String[] constantTags, final StatsDClientErrorHandler errorHandler,
+                                final Callable<InetSocketAddress> addressLookup) throws StatsDClientException {
+        if((prefix != null) && (!prefix.isEmpty())) {
+            this.prefix = String.format("%s.", prefix);
+        } else {
+            this.prefix = "";
+        }
+        if(errorHandler == null) {
+            handler = StatsDClientErrorHandler.NO_OP_HANDLER;
+        }
+        else {
+            handler = errorHandler;
+        }
+
+        /* Empty list should be null for faster comparison */
+        if((constantTags != null) && (constantTags.length == 0)) {
+            constantTags = null;
+        }
+
+        if(constantTags != null) {
+            constantTagsRendered = tagString(constantTags, null);
+        } else {
+            constantTagsRendered = null;
+        }
+
+        try {
+            clientChannel = DatagramChannel.open();
+        } catch (final Exception e) {
+            throw new StatsDClientException("Failed to start StatsD client", e);
+        }
+        this.addressLookup = addressLookup;
+    }
+
+    /**
+     * Cleanly shut down this StatsD client. This method may throw an exception if
+     * the socket cannot be closed.
+     */
+    @Override
+    public void stop() {
+        if (clientChannel != null) {
+            try {
+                clientChannel.close();
+            }
+            catch (IOException e) {
+                handler.handle(e);
+            }
+        }
+    }
+
+    /**
+     * Generate a suffix conveying the given tag list to the client
+     */
+    static String tagString(final String[] tags, final String tagPrefix) {
+        final StringBuilder sb;
+        if(tagPrefix != null) {
+            if((tags == null) || (tags.length == 0)) {
+                return tagPrefix;
+            }
+            sb = new StringBuilder(tagPrefix);
+            sb.append(",");
+        } else {
+            if((tags == null) || (tags.length == 0)) {
+                return "";
+            }
+            sb = new StringBuilder("|#");
+        }
+
+        for(int n=tags.length - 1; n>=0; n--) {
+            sb.append(tags[n]);
+            if(n > 0) {
+                sb.append(",");
+            }
+        }
+        return sb.toString();
+    }
+
+    /**
+     * Generate a suffix conveying the given tag list to the client
+     */
+    String tagString(final String[] tags) {
+        return tagString(tags, constantTagsRendered);
+    }
+
+    /**
+     * Adjusts the specified counter by a given delta.
+     *
+     * <p>This method is non-blocking and is guaranteed not to throw an exception.</p>
+     *
+     * @param aspect
+     *     the name of the counter to adjust
+     * @param delta
+     *     the amount to adjust the counter by
+     * @param tags
+     *     array of tags to be added to the data
+     */
+    @Override
+    public void count(final String aspect, final long delta, final String... tags) {
+        send(String.format("%s%s:%d|c%s", prefix, aspect, delta, tagString(tags)));
+    }
+
+    /**
+     * Increments the specified counter by one.
+     *
+     * <p>This method is non-blocking and is guaranteed not to throw an exception.</p>
+     *
+     * @param aspect
+     *     the name of the counter to increment
+     * @param tags
+     *     array of tags to be added to the data
+     */
+    @Override
+    public void incrementCounter(final String aspect, final String... tags) {
+        count(aspect, 1, tags);
+    }
+
+    /**
+     * Convenience method equivalent to {@link #incrementCounter(String, String[])}.
+     */
+    @Override
+    public void increment(final String aspect, final String... tags) {
+        incrementCounter(aspect, tags);
+    }
+
+    /**
+     * Decrements the specified counter by one.
+     *
+     * <p>This method is non-blocking and is guaranteed not to throw an exception.</p>
+     *
+     * @param aspect
+     *     the name of the counter to decrement
+     * @param tags
+     *     array of tags to be added to the data
+     */
+    @Override
+    public void decrementCounter(final String aspect, final String... tags) {
+        count(aspect, -1, tags);
+    }
+
+    /**
+     * Convenience method equivalent to {@link #decrementCounter(String, String[])}.
+     */
+    @Override
+    public void decrement(final String aspect, final String... tags) {
+        decrementCounter(aspect, tags);
+    }
+
+    /**
+     * Records the latest fixed value for the specified named gauge.
+     *
+     * <p>This method is non-blocking and is guaranteed not to throw an exception.</p>
+     *
+     * @param aspect
+     *     the name of the gauge
+     * @param value
+     *     the new reading of the gauge
+     * @param tags
+     *     array of tags to be added to the data
+     */
+    @Override
+    public void recordGaugeValue(final String aspect, final double value, final String... tags) {
+        /* Intentionally using %s rather than %f here to avoid
+         * padding with extra 0s to represent precision */
+        send(String.format("%s%s:%s|g%s", prefix, aspect, NUMBER_FORMATTERS.get().format(value), tagString(tags)));
+    }
+
+    /**
+     * Convenience method equivalent to {@link #recordGaugeValue(String, double, String[])}.
+     */
+    @Override
+    public void gauge(final String aspect, final double value, final String... tags) {
+        recordGaugeValue(aspect, value, tags);
+    }
+
+
+    /**
+     * Records the latest fixed value for the specified named gauge.
+     *
+     * <p>This method is non-blocking and is guaranteed not to throw an exception.</p>
+     *
+     * @param aspect
+     *     the name of the gauge
+     * @param value
+     *     the new reading of the gauge
+     * @param tags
+     *     array of tags to be added to the data
+     */
+    @Override
+    public void recordGaugeValue(final String aspect, final long value, final String... tags) {
+        send(String.format("%s%s:%d|g%s", prefix, aspect, value, tagString(tags)));
+    }
+
+    /**
+     * Convenience method equivalent to {@link #recordGaugeValue(String, long, String[])}.
+     */
+    @Override
+    public void gauge(final String aspect, final long value, final String... tags) {
+        recordGaugeValue(aspect, value, tags);
+    }
+
+    /**
+     * Records an execution time in milliseconds for the specified named operation.
+     *
+     * <p>This method is non-blocking and is guaranteed not to throw an exception.</p>
+     *
+     * @param aspect
+     *     the name of the timed operation
+     * @param timeInMs
+     *     the time in milliseconds
+     * @param tags
+     *     array of tags to be added to the data
+     */
+    @Override
+    public void recordExecutionTime(final String aspect, final long timeInMs, final String... tags) {
+        send(String.format("%s%s:%d|ms%s", prefix, aspect, timeInMs, tagString(tags)));
+    }
+
+    /**
+     * Convenience method equivalent to {@link #recordExecutionTime(String, long, String[])}.
+     */
+    @Override
+    public void time(final String aspect, final long value, final String... tags) {
+        recordExecutionTime(aspect, value, tags);
+    }
+
+    /**
+     * Records a value for the specified named histogram.
+     *
+     * <p>This method is non-blocking and is guaranteed not to throw an exception.</p>
+     *
+     * @param aspect
+     *     the name of the histogram
+     * @param value
+     *     the value to be incorporated in the histogram
+     * @param tags
+     *     array of tags to be added to the data
+     */
+    @Override
+    public void recordHistogramValue(final String aspect, final double value, final String... tags) {
+        /* Intentionally using %s rather than %f here to avoid
+         * padding with extra 0s to represent precision */
+        send(String.format("%s%s:%s|h%s", prefix, aspect, NUMBER_FORMATTERS.get().format(value), tagString(tags)));
+    }
+
+    /**
+     * Convenience method equivalent to {@link #recordHistogramValue(String, double, String[])}.
+     */
+    @Override
+    public void histogram(final String aspect, final double value, final String... tags) {
+        recordHistogramValue(aspect, value, tags);
+    }
+
+    /**
+     * Records a value for the specified named histogram.
+     *
+     * <p>This method is non-blocking and is guaranteed not to throw an exception.</p>
+     *
+     * @param aspect
+     *     the name of the histogram
+     * @param value
+     *     the value to be incorporated in the histogram
+     * @param tags
+     *     array of tags to be added to the data
+     */
+    @Override
+    public void recordHistogramValue(final String aspect, final long value, final String... tags) {
+        send(String.format("%s%s:%d|h%s", prefix, aspect, value, tagString(tags)));
+    }
+
+    /**
+     * Convenience method equivalent to {@link #recordHistogramValue(String, long, String[])}.
+     */
+    @Override
+    public void histogram(final String aspect, final long value, final String... tags) {
+        recordHistogramValue(aspect, value, tags);
+    }
+
+    private String eventMap(final Event event) {
+        final StringBuilder res = new StringBuilder("");
+
+        final long millisSinceEpoch = event.getMillisSinceEpoch();
+        if (millisSinceEpoch != -1) {
+            res.append("|d:").append(millisSinceEpoch / 1000);
+        }
+
+        final String hostname = event.getHostname();
+        if (hostname != null) {
+            res.append("|h:").append(hostname);
+        }
+
+        final String aggregationKey = event.getAggregationKey();
+        if (aggregationKey != null) {
+            res.append("|k:").append(aggregationKey);
+        }
+
+        final String priority = event.getPriority();
+        if (priority != null) {
+            res.append("|p:").append(priority);
+        }
+
+        final String alertType = event.getAlertType();
+        if (alertType != null) {
+            res.append("|t:").append(alertType);
+        }
+
+        return res.toString();
+    }
+
+    /**
+     * Records an event
+     *
+     * <p>This method is a DataDog extension, and may not work with other servers.</p>
+     *
+     * <p>This method is non-blocking and is guaranteed not to throw an exception.</p>
+     *
+     * @param event
+     *     The event to record
+     * @param tags
+     *     array of tags to be added to the data
+     *
+     * @see <a href="http://docs.datadoghq.com/guides/dogstatsd/#events-1">http://docs.datadoghq.com/guides/dogstatsd/#events-1</a>
+     */
+    @Override
+    public void recordEvent(final Event event, final String... tags) {
+        final String title = prefix + event.getTitle();
+        final String text = event.getText();
+        send(String.format("_e{%d,%d}:%s|%s%s%s",
+                title.length(), text.length(), title, text, eventMap(event), tagString(tags)));
+    }
+
+    /**
+     * Records a run status for the specified named service check.
+     *
+     * <p>This method is a DataDog extension, and may not work with other servers.</p>
+     *
+     * <p>This method is non-blocking and is guaranteed not to throw an exception.</p>
+     *
+     * @param sc
+     *     the service check object
+     */
+    @Override
+    public void recordServiceCheckRun(final ServiceCheck sc) {
+        send(toStatsDString(sc));
+    }
+    
+    private String toStatsDString(final ServiceCheck sc) {
+        // see http://docs.datadoghq.com/guides/dogstatsd/#service-checks
+        final StringBuilder sb = new StringBuilder();
+        sb.append(String.format("_sc|%s|%d", sc.getName(), sc.getStatus()));
+        if (sc.getTimestamp() > 0) {
+            sb.append(String.format("|d:%d", sc.getTimestamp()));
+        }
+        if (sc.getHostname() != null) {
+            sb.append(String.format("|h:%s", sc.getHostname()));
+        }
+        sb.append(tagString(sc.getTags()));
+        if (sc.getMessage() != null) {
+            sb.append(String.format("|m:%s", sc.getEscapedMessage()));
+        }
+        return sb.toString();
+    }
+
+    /**
+     * Convenience method equivalent to {@link #recordServiceCheckRun(ServiceCheck sc)}.
+     */
+    @Override
+    public void serviceCheck(final ServiceCheck sc) {
+        recordServiceCheckRun(sc);
+    }
+
+
+    /**
+     * Records a value for the specified set.
+     *
+     * Sets are used to count the number of unique elements in a group. If you want to track the number of
+     * unique visitor to your site, sets are a great way to do that.
+     *
+     * <p>This method is a DataDog extension, and may not work with other servers.</p>
+     *
+     * <p>This method is non-blocking and is guaranteed not to throw an exception.</p>
+     *
+     * @param aspect
+     *     the name of the set
+     * @param value
+     *     the value to track
+     * @param tags
+     *     array of tags to be added to the data
+     *
+     * @see <a href="http://docs.datadoghq.com/guides/dogstatsd/#sets">http://docs.datadoghq.com/guides/dogstatsd/#sets</a>
+     */
+    @Override
+    public void recordSetValue(final String aspect, final String value, final String... tags) {
+        // documentation is light, but looking at dogstatsd source, we can send string values
+        // here instead of numbers
+        send(String.format("%s%s:%s|s%s", prefix, aspect, value, tagString(tags)));
+    }
+
+
+    public static final Charset MESSAGE_CHARSET = Charset.forName("UTF-8");
+    private final ByteBuffer sendBuffer = ByteBuffer.allocate(PACKET_SIZE_BYTES);
+
+    protected void send(String message) {
+        if (message == null) return;
+        
+        try {
+            final InetSocketAddress address = addressLookup.call();
+            final byte[] data = message.getBytes(MESSAGE_CHARSET);
+
+            sendBuffer.put(data);
+
+            final int sizeOfBuffer = sendBuffer.position();
+            sendBuffer.flip();
+
+            final int sentBytes = clientChannel.send(sendBuffer, address);
+            sendBuffer.limit(sendBuffer.capacity());
+            sendBuffer.rewind();
+
+            if (sizeOfBuffer != sentBytes) {
+                handler.handle(
+                        new IOException(
+                            String.format(
+                                "Could not send entirely stat %s to host %s:%d. Only sent %d bytes out of %d bytes",
+                                sendBuffer.toString(),
+                                address.getHostName(),
+                                address.getPort(),
+                                sentBytes,
+                                sizeOfBuffer)));
+            }
+        } catch (Exception e) {
+            handler.handle(e);
+        }
+    }
+
+    /**
+     * Create dynamic lookup for the given host name and port.
+     *
+     * @param hostname the host name of the targeted StatsD server
+     * @param port     the port of the targeted StatsD server
+     * @return a function to perform the lookup
+     * @see BlockingStatsDClient#BlockingStatsDClient(String, String[], StatsDClientErrorHandler, Callable)
+     */
+    public static Callable<InetSocketAddress> volatileAddressResolution(final String hostname, final int port) {
+        return new Callable<InetSocketAddress>() {
+            @Override public InetSocketAddress call() throws UnknownHostException {
+                return new InetSocketAddress(InetAddress.getByName(hostname), port);
+            }
+        };
+    }
+
+  /**
+   * Lookup the address for the given host name and cache the result.
+   *
+   * @param hostname the host name of the targeted StatsD server
+   * @param port     the port of the targeted StatsD server
+   * @return a function that cached the result of the lookup
+   * @throws Exception if the lookup fails, i.e. {@link UnknownHostException}
+   */
+    public static Callable<InetSocketAddress> staticAddressResolution(final String hostname, final int port) throws Exception {
+        final InetSocketAddress address = volatileAddressResolution(hostname, port).call();
+        return new Callable<InetSocketAddress>() {
+            @Override public InetSocketAddress call() {
+                return address;
+            }
+        };
+    }
+
+    protected static Callable<InetSocketAddress> staticStatsDAddressResolution(final String hostname, final int port) throws StatsDClientException {
+        try {
+            return staticAddressResolution(hostname, port);
+        } catch (final Exception e) {
+            throw new StatsDClientException("Failed to lookup StatsD host", e);
+        }
+    }
+}
