+++ src/test/java/com/timgroup/statsd/StatsDClientTest.java
@@ -1,0 +1,441 @@
+package com.timgroup.statsd;
+
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+
+import java.net.SocketException;
+import java.util.Locale;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.contains;
+import static org.junit.Assert.assertEquals;
+
+public abstract class StatsDClientTest<T extends StatsDClient> {
+
+    protected static final String STATSD_SERVER_HOST = "localhost";
+    protected static final int STATSD_SERVER_PORT = 17254;
+    
+    private DummyStatsDServer server;
+    
+    public T buildClient(String prefix) {
+        return buildClient(prefix, null);
+    }
+    
+    abstract public T buildClient(String prefix, String[] constantTags);
+
+    private final T client = buildClient("my.prefix");
+
+    @Before
+    public void start() throws SocketException {
+        server = new DummyStatsDServer(STATSD_SERVER_PORT);
+    }
+
+    @After
+    public void stop() throws Exception {
+        client.stop();
+        server.close();
+    }
+
+    @Test(timeout=5000L) public void
+    sends_counter_value_to_statsd() throws Exception {
+
+
+        client.count("mycount", 24);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.mycount:24|c"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_counter_value_to_statsd_with_null_tags() throws Exception {
+
+
+        client.count("mycount", 24, (java.lang.String[]) null);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.mycount:24|c"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_counter_value_to_statsd_with_empty_tags() throws Exception {
+
+
+        client.count("mycount", 24);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.mycount:24|c"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_counter_value_to_statsd_with_tags() throws Exception {
+
+
+        client.count("mycount", 24, "foo:bar", "baz");
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.mycount:24|c|#baz,foo:bar"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_counter_increment_to_statsd() throws Exception {
+
+
+        client.incrementCounter("myinc");
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.myinc:1|c"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_counter_increment_to_statsd_with_tags() throws Exception {
+
+
+        client.incrementCounter("myinc", "foo:bar", "baz");
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.myinc:1|c|#baz,foo:bar"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_counter_decrement_to_statsd() throws Exception {
+
+
+        client.decrementCounter("mydec");
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.mydec:-1|c"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_counter_decrement_to_statsd_with_tags() throws Exception {
+
+
+        client.decrementCounter("mydec", "foo:bar", "baz");
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.mydec:-1|c|#baz,foo:bar"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_gauge_to_statsd() throws Exception {
+
+
+        client.recordGaugeValue("mygauge", 423);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.mygauge:423|g"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_large_double_gauge_to_statsd() throws Exception {
+
+
+        client.recordGaugeValue("mygauge", 123456789012345.67890);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.mygauge:123456789012345.67|g"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_exact_double_gauge_to_statsd() throws Exception {
+
+
+        client.recordGaugeValue("mygauge", 123.45678901234567890);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.mygauge:123.456789|g"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_double_gauge_to_statsd() throws Exception {
+
+
+        client.recordGaugeValue("mygauge", 0.423);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.mygauge:0.423|g"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_gauge_to_statsd_with_tags() throws Exception {
+
+
+        client.recordGaugeValue("mygauge", 423, "foo:bar", "baz");
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.mygauge:423|g|#baz,foo:bar"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_double_gauge_to_statsd_with_tags() throws Exception {
+
+
+        client.recordGaugeValue("mygauge", 0.423, "foo:bar", "baz");
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.mygauge:0.423|g|#baz,foo:bar"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_histogram_to_statsd() throws Exception {
+
+
+        client.recordHistogramValue("myhistogram", 423);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.myhistogram:423|h"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_double_histogram_to_statsd() throws Exception {
+
+
+        client.recordHistogramValue("myhistogram", 0.423);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.myhistogram:0.423|h"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_histogram_to_statsd_with_tags() throws Exception {
+
+
+        client.recordHistogramValue("myhistogram", 423, "foo:bar", "baz");
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.myhistogram:423|h|#baz,foo:bar"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_double_histogram_to_statsd_with_tags() throws Exception {
+
+
+        client.recordHistogramValue("myhistogram", 0.423, "foo:bar", "baz");
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.myhistogram:0.423|h|#baz,foo:bar"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_timer_to_statsd() throws Exception {
+
+
+        client.recordExecutionTime("mytime", 123);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.mytime:123|ms"));
+    }
+
+    /**
+     * A regression test for <a href="https://github.com/indeedeng/java-dogstatsd-client/issues/3">this i18n number formatting bug</a>
+     * @throws Exception
+     */
+    @Test public void
+    sends_timer_to_statsd_from_locale_with_unamerican_number_formatting() throws Exception {
+
+        Locale originalDefaultLocale = Locale.getDefault();
+
+        // change the default Locale to one that uses something other than a '.' as the decimal separator (Germany uses a comma)
+        Locale.setDefault(Locale.GERMANY);
+
+        try {
+
+
+            client.recordExecutionTime("mytime", 123, "foo:bar", "baz");
+            server.waitForMessage();
+
+            assertThat(server.messagesReceived(), contains("my.prefix.mytime:123|ms|#baz,foo:bar"));
+        } finally {
+            // reset the default Locale in case changing it has side-effects
+            Locale.setDefault(originalDefaultLocale);
+        }
+    }
+
+
+    @Test(timeout=5000L) public void
+    sends_timer_to_statsd_with_tags() throws Exception {
+
+
+        client.recordExecutionTime("mytime", 123, "foo:bar", "baz");
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.mytime:123|ms|#baz,foo:bar"));
+    }
+
+
+    @Test(timeout=5000L) public void
+    sends_gauge_mixed_tags() throws Exception {
+
+        final T empty_prefix_client = buildClient("my.prefix", new String[] {"instance:foo", "app:bar"});
+        empty_prefix_client.gauge("value", 423, "baz");
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.value:423|g|#app:bar,instance:foo,baz"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_gauge_constant_tags_only() throws Exception {
+
+        final T empty_prefix_client = buildClient("my.prefix", new String[] {"instance:foo", "app:bar"});
+        empty_prefix_client.gauge("value", 423);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.value:423|g|#app:bar,instance:foo"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_gauge_empty_prefix() throws Exception {
+
+        final T empty_prefix_client = buildClient("");
+        empty_prefix_client.gauge("top.level.value", 423);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("top.level.value:423|g"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_gauge_null_prefix() throws Exception {
+
+        final T null_prefix_client = buildClient(null);
+        null_prefix_client.gauge("top.level.value", 423);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("top.level.value:423|g"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_event() throws Exception {
+
+        final Event event = Event.builder()
+                .withTitle("title1")
+                .withText("text1")
+                .withDate(1234567000)
+                .withHostname("host1")
+                .withPriority(Event.Priority.LOW)
+                .withAggregationKey("key1")
+                .withAlertType(Event.AlertType.ERROR)
+                .build();
+        client.recordEvent(event);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("_e{16,5}:my.prefix.title1|text1|d:1234567|h:host1|k:key1|p:low|t:error"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_partial_event() throws Exception {
+
+        final Event event = Event.builder()
+                .withTitle("title1")
+                .withText("text1")
+                .withDate(1234567000)
+                .build();
+        client.recordEvent(event);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("_e{16,5}:my.prefix.title1|text1|d:1234567"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_event_with_tags() throws Exception {
+
+        final Event event = Event.builder()
+                .withTitle("title1")
+                .withText("text1")
+                .withDate(1234567000)
+                .withHostname("host1")
+                .withPriority(Event.Priority.LOW)
+                .withAggregationKey("key1")
+                .withAlertType(Event.AlertType.ERROR)
+                .build();
+        client.recordEvent(event, "foo:bar", "baz");
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("_e{16,5}:my.prefix.title1|text1|d:1234567|h:host1|k:key1|p:low|t:error|#baz,foo:bar"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_partial_event_with_tags() throws Exception {
+
+        final Event event = Event.builder()
+                .withTitle("title1")
+                .withText("text1")
+                .withDate(1234567000)
+                .build();
+        client.recordEvent(event, "foo:bar", "baz");
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("_e{16,5}:my.prefix.title1|text1|d:1234567|#baz,foo:bar"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_event_empty_prefix() throws Exception {
+
+        final T empty_prefix_client = buildClient("");
+        final Event event = Event.builder()
+                .withTitle("title1")
+                .withText("text1")
+                .withDate(1234567000)
+                .withHostname("host1")
+                .withPriority(Event.Priority.LOW)
+                .withAggregationKey("key1")
+                .withAlertType(Event.AlertType.ERROR)
+                .build();
+        empty_prefix_client.recordEvent(event, "foo:bar", "baz");
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("_e{6,5}:title1|text1|d:1234567|h:host1|k:key1|p:low|t:error|#baz,foo:bar"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_service_check() throws Exception {
+        final String inputMessage = "\u266c \u2020\u00f8U \n\u2020\u00f8U \u00a5\u00bau|m: T0\u00b5 \u266a"; // "♬ †øU \n†øU ¥ºu|m: T0µ ♪"
+        final String outputMessage = "\u266c \u2020\u00f8U \\n\u2020\u00f8U \u00a5\u00bau|m\\: T0\u00b5 \u266a"; // note the escaped colon
+        final String[] tags = {"key1:val1", "key2:val2"};
+        final ServiceCheck sc = ServiceCheck.builder()
+                .withName("my_check.name")
+                .withStatus(ServiceCheck.Status.WARNING)
+                .withMessage(inputMessage)
+                .withHostname("i-abcd1234")
+                .withTags(tags)
+                .withTimestamp(1420740000)
+                .build();
+
+        assertEquals(outputMessage, sc.getEscapedMessage());
+
+        client.serviceCheck(sc);
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains(String.format("_sc|my_check.name|1|d:1420740000|h:i-abcd1234|#key2:val2,key1:val1|m:%s",
+                outputMessage)));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_nan_gauge_to_statsd() throws Exception {
+        client.recordGaugeValue("mygauge", Double.NaN);
+
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.mygauge:NaN|g"));
+    }
+
+    @Test(timeout=5000L) public void
+    sends_set_to_statsd() throws Exception {
+        client.recordSetValue("myset", "myuserid");
+
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.myset:myuserid|s"));
+
+    }
+
+    @Test(timeout=5000L) public void
+    sends_set_to_statsd_with_tags() throws Exception {
+        client.recordSetValue("myset", "myuserid", "foo:bar", "baz");
+
+        server.waitForMessage();
+
+        assertThat(server.messagesReceived(), contains("my.prefix.myset:myuserid|s|#baz,foo:bar"));
+
+    }
+}
